requirements-merged.txt
Django==5.0.6
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
ultralytics==8.3.0
opencv-python==4.10.0.84
torch==2.3.0
tensorflow==2.16.1
